[TEST] My Keyword Total time 테스트 보완
