+++ src/tests/E2E/Mypage.test.tsx
@@ -1,9 +1,11 @@
 import React from "react";
-import { screen, waitFor } from "@testing-library/react";
+import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
 import { renderLoggedInLayout, renderLoggedOutLayout } from "@/utils/util-test";
 import MyPagePage from "@/app/mypage/page";
+import MyPageCategory from "@/components/mypage/article/MyPageCategory";
+import Provider from "@/components/provider/Provider";
 
 import { mockNavigation } from "../__mocks__/next/navigation";
 
@@ -16,7 +18,7 @@ const renderMypage = {
   },
 };
 
-describe("마이페이지 화면 테스트", () => {
+describe("렌더링 테스트", () => {
   test("비로그인일 경우 접근시 로그인 페이지로 이동", async () => {
     renderMypage.loggedOut();
     await waitFor(() => {
@@ -28,19 +30,28 @@ describe("마이페이지 화면 테스트", () => {
     renderMypage.loggedIn();
     expect(await screen.findByTestId("renderedMypage")).toBeInTheDocument();
   });
+});
 
-  test("검색창에 내 카테고리를 검색하면 자동추천이 노출된다.", async () => {
-    renderMypage.loggedIn();
-    const searchInput = await screen.findByPlaceholderText(
-      "내가 추가한 나만의 키워드를 검색해보세요.",
-    );
-    await waitFor(() => {
-      expect(screen.getByText("# 채팅방입니다1")).toBeInTheDocument();
-    });
-    await userEvent.type(searchInput, "채팅방");
-    screen.getByText("채팅방입니다1");
+test("My Keyword Total time 검색창에 내 카테고리를 검색하면 자동추천이 노출되고 선택하면 리스트가 필터링되어야 함.", async () => {
+  render(<MyPageCategory />, { wrapper: Provider });
+  const searchInput = await screen.findByPlaceholderText(
+    "내가 추가한 나만의 키워드를 검색해보세요.",
+  );
+
+  await waitFor(() => {
+    expect(screen.getByText("# 리액트")).toBeInTheDocument();
+    expect(screen.getByText("# Javascript")).toBeInTheDocument();
   });
+  await userEvent.type(searchInput, "Javascript");
+  expect(screen.queryByText("리액트")).not.toBeInTheDocument();
+  await userEvent.clear(searchInput);
+  await userEvent.type(searchInput, "리");
+  await userEvent.click(await screen.findByText("리액트"));
+  expect(screen.queryByText("# Javascript")).not.toBeInTheDocument();
+  expect(await screen.findByText("# 리액트")).toBeInTheDocument();
+});
 
+describe("테스트", () => {
   test("프로필 이미지 수정 기능", async () => {
     renderMypage.loggedIn();
     const file = new File(["testImg"], "testImg.png", {
